// EventFrontend2.O/src/app/core/models/models.ts
export interface pageSelection {
  skip: number;
  limit: number;
}
export interface apiResultFormat {
  // eslint-disable-next-line @typescript-eslint/no-explicit-any
  data: any;
  totalData: number;
}
export interface vendor {
  id: number;
  sNo: number;
  name: string;
  name_urdu: string;
  img: string;
  email: string;
  phone: string;
  created: string;
  balance: string;
  category: any;
  subcategory: string;
}
// ...
export interface category {
  sNo: number;
  date: string | number;
  img2: string;
  added: string;
  id: number;
  category: string;
  img: string;
  total: number;
  value: string;
  description: string;
}